+++ src/Main/Setup.cpp
@@ -40,13 +40,12 @@ Setup::Setup()
 
   // count of runs in conjunction of calls of altersetup
   runcount = 0;
-
   write_meta_file=false;
   sema_file_enabled_start=false;
   sema_file_enabled_done=false;
 }
 
-Setup::~Setup(){}
+Setup::~Setup()= default;
 
 void Setup::usage(){
 
@@ -79,13 +78,12 @@ void Setup::usage(){
   cout << " bool write_semaphore_file_started = false" << endl;
   cout << " string semaphore_file_name = <derived from 'rootname'>" << endl;
   cout << "&end" << endl << endl;
-  return;
 }
 
 bool Setup::init(int inrank, map<string,string> *arg, Lattice *lat, SeriesManager *sm, FilterDiagnostics &filter){
 
   rank=inrank;
-  map<string,string>::iterator end=arg->end();
+  auto end=arg->end();
 
   if (arg->find("rootname")!=end){rootname = arg->at("rootname"); arg->erase(arg->find("rootname"));}
   if (arg->find("outputdir")!=end){outputdir = arg->at("outputdir"); arg->erase(arg->find("outputdir"));}
@@ -145,7 +143,7 @@ bool Setup::init(int inrank, map<string,string> *arg, Lattice *lat, SeriesManage
     arg->erase(arg->find("beam_write_slices_filter"));
   }
 
-  if (arg->size()!=0){
+  if (!arg->empty()){
     if (rank==0){ cout << "*** Error: Unknown elements in &setup" << endl; this->usage();}
     return false;
   }
@@ -173,7 +171,7 @@ bool Setup::init(int inrank, map<string,string> *arg, Lattice *lat, SeriesManage
 
 bool Setup::getRootName(string *filename)
 {
-  if (rootname.size()<1){
+  if (rootname.empty()){
     return false;
   }
   *filename=rootname;
@@ -184,17 +182,25 @@ bool Setup::getRootName(string *filename)
   }
   return true; 
 }
-bool Setup::RootName_to_FileName(string *fn_out, string *rootname)
+
+bool Setup::RootName_to_FileName(string *fn_out, string *filename)
 {
+    // replace placeholder symbol @ with rootname
+    std::string placeholder("@");
+    int pos;
+    while ((pos = filename->find(placeholder)) != std::string::npos)
+        filename->replace(pos, placeholder.length(),rootname);
+
   // If 'outputdir' parameter not specified
   // ==> do not change string
-  if (outputdir.size()<1) {
-    *fn_out = *rootname;
+
+  if (outputdir.empty()) {
+    *fn_out = *filename;
     return true;
   }
 
   string t;
-  t = outputdir + "/" + *rootname;
+  t = outputdir + "/" + *filename;
   *fn_out = t;
 
   return true;
@@ -202,7 +208,7 @@ bool Setup::RootName_to_FileName(string *fn_out, string *rootname)
 
 void Setup::BWF_load_defaults()
 {
-  // note: only loading defaults, not diabling the filter (if desired, this has to be done in addition)
+  // note: only loading defaults, not disabling the filter (if desired, this has to be done in addition)
   beam_write_slices_from=-1;
   beam_write_slices_to=-1;
   beam_write_slices_inc=1;