+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 # Change Log
+### [4.6.6] 
+- nothing yet
 
 ### [4.6.5] - 2023/12/06
 - added the ability to reduce the amount of particles when writing to a beam dump